ceph: refactor csi related code into a new file in csi package
